Move MavenProject to parent package
