+++ modules/models/java-project/src/main/kotlin/org/cafejojo/schaapi/models/project/java/JavaMavenProject.kt
@@ -1,5 +1,6 @@
 package org.cafejojo.schaapi.models.project.java
 
+import org.cafejojo.schaapi.models.project.MavenProject
 import java.io.File
 
 /**
